requirements.txt
absl-py==2.1.0
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
aiosqlite==0.21.0
alembic==1.14.1
altair==4.1.0
anyio==4.8.0
apache-airflow==2.10.5
apache-airflow-providers-common-compat==1.3.0
apache-airflow-providers-common-io==1.5.0
apache-airflow-providers-common-sql==1.21.0
apache-airflow-providers-fab==1.5.3
apache-airflow-providers-ftp==3.12.0
apache-airflow-providers-http==5.0.0
apache-airflow-providers-imap==3.8.0
apache-airflow-providers-smtp==1.9.0
apache-airflow-providers-sqlite==4.0.0
apispec==6.8.1
argcomplete==3.5.3
asgiref==3.8.1
asttokens @ file:///home/conda/feedstock_root/build_artifacts/asttokens_1698341106958/work
astunparse==1.6.3
async-timeout==5.0.1
attrs @ file:///home/conda/feedstock_root/build_artifacts/attrs_1722977137225/work
Automat==24.8.1
babel==2.17.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachelib==0.9.0
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
clickclick==20.10.2
colorama @ file:///home/conda/feedstock_root/build_artifacts/colorama_1666700638685/work
colorlog==6.9.0
comm @ file:///home/conda/feedstock_root/build_artifacts/comm_1710320294760/work
ConfigUpdater==3.2
connexion==2.14.2
constantly==23.10.4
contourpy==1.3.0
cron-descriptor==1.4.5
croniter==6.0.0
cryptography==44.0.1
cssselect==1.2.0
cycler==0.12.1
debugpy @ file:///D:/bld/debugpy_1732236679090/work
decorator @ file:///home/conda/feedstock_root/build_artifacts/decorator_1641555617451/work
defusedxml==0.7.1
Deprecated==1.2.18
dill==0.3.9
dnspython==2.7.0
email_validator==2.2.0
entrypoints==0.4
eval_type_backport==0.2.2
exceptiongroup @ file:///home/conda/feedstock_root/build_artifacts/exceptiongroup_1720869315914/work
executing @ file:///home/conda/feedstock_root/build_artifacts/executing_1725214404607/work
filelock==3.17.0
Flask==2.2.5
Flask-AppBuilder==4.5.3
Flask-Babel==2.0.0
Flask-Caching==2.3.0
Flask-JWT-Extended==4.7.1
Flask-Limiter==3.10.1
Flask-Login==0.6.3
Flask-Session==0.5.0
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.2.2
flatbuffers==24.3.25
fonttools==4.55.0
frozenlist==1.5.0
fsspec==2025.2.0
gast==0.4.0
geographiclib==2.0
geopy==2.4.1
gitdb==4.0.11
GitPython==3.1.43
google-auth==2.36.0
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
google-re2==1.1.20240702
googleapis-common-protos==1.67.0
greenlet==3.1.1
grpcio==1.68.0
gunicorn==23.0.0
h11==0.14.0
h5py==3.12.1
httpcore==1.0.7
httpx==0.28.1
hyperlink==21.0.0
idna==3.10
imageio==2.36.1
imgaug==0.4.0
importlib_metadata @ file:///home/conda/feedstock_root/build_artifacts/importlib-metadata_1726082825846/work
importlib_resources @ file:///home/conda/feedstock_root/build_artifacts/importlib_resources_1733231327252/work
incremental==24.7.2
inflection==0.5.1
ipykernel @ file:///D:/bld/ipykernel_1719845595208/work
ipython @ file:///D:/bld/ipython_1701831845989/work
itemadapter==0.11.0
itemloaders==1.3.2
itsdangerous==2.2.0
jedi @ file:///home/conda/feedstock_root/build_artifacts/jedi_1731317204262/work
Jinja2 @ file:///home/conda/feedstock_root/build_artifacts/jinja2_1733217336947/work
jmespath==1.0.1
joblib==1.4.2
jsonschema @ file:///home/conda/feedstock_root/build_artifacts/jsonschema_1733472696581/work
jsonschema-specifications @ file:///tmp/tmpk0f344m9/src
jupyter_client @ file:///home/conda/feedstock_root/build_artifacts/jupyter_client_1726610684920/work
jupyter_core @ file:///D:/bld/jupyter_core_1727163532151/work
keras==2.10.0
Keras-Preprocessing==1.1.2
kiwisolver==1.4.7
lazy-object-proxy==1.10.0
lazy_loader==0.4
libclang==18.1.1
limits==4.0.1
linkify-it-py==2.0.3
lockfile==0.12.2
lxml==5.3.1
Mako==1.3.9
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe @ file:///D:/bld/markupsafe_1733219648915/work
marshmallow==3.26.1
marshmallow-oneofschema==3.1.1
marshmallow-sqlalchemy==0.28.2
matplotlib==3.9.3
matplotlib-inline @ file:///home/conda/feedstock_root/build_artifacts/matplotlib-inline_1713250518406/work
mdit-py-plugins==0.4.2
mdurl==0.1.2
methodtools==0.4.7
ml-dtypes==0.4.1
more-itertools==10.6.0
multidict==6.1.0
namex==0.0.8
narwhals @ file:///home/conda/feedstock_root/build_artifacts/narwhals_1733267207975/work
nest_asyncio @ file:///home/conda/feedstock_root/build_artifacts/nest-asyncio_1705850609492/work
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
opencv-python==4.10.0.84
opentelemetry-api==1.30.0
opentelemetry-exporter-otlp==1.30.0
opentelemetry-exporter-otlp-proto-common==1.30.0
opentelemetry-exporter-otlp-proto-grpc==1.30.0
opentelemetry-exporter-otlp-proto-http==1.30.0
opentelemetry-proto==1.30.0
opentelemetry-sdk==1.30.0
opentelemetry-semantic-conventions==0.51b0
opt_einsum==3.4.0
optree==0.13.1
ordered-set==4.1.0
packaging @ file:///home/conda/feedstock_root/build_artifacts/packaging_1731802491770/work
pandas==2.2.3
parsel==1.10.0
parso @ file:///home/conda/feedstock_root/build_artifacts/parso_1712320355065/work
pathspec==0.12.1
pendulum==3.0.0
pickleshare @ file:///home/conda/feedstock_root/build_artifacts/pickleshare_1602536217715/work
pillow==11.0.0
pkgutil_resolve_name @ file:///home/conda/feedstock_root/build_artifacts/pkgutil-resolve-name_1733344503739/work
platformdirs @ file:///home/conda/feedstock_root/build_artifacts/platformdirs_1726613481435/work
pluggy==1.5.0
prison==0.2.1
prompt_toolkit @ file:///home/conda/feedstock_root/build_artifacts/prompt-toolkit_1727341649933/work
propcache==0.2.1
Protego==0.4.0
protobuf==5.29.3
psutil @ file:///D:/bld/psutil_1729847171148/work
psycopg2==2.9.10
pure_eval @ file:///home/conda/feedstock_root/build_artifacts/pure_eval_1721585709575/work
py4j==0.10.9.7
pyarrow==18.1.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pydeck==0.9.1
PyDispatcher==2.0.7
Pygments @ file:///home/conda/feedstock_root/build_artifacts/pygments_1714846767233/work
PyJWT==2.10.1
Pympler==1.1
pyOpenSSL==25.0.0
pyparsing==3.2.0
pyspark==3.5.4
python-daemon==3.1.2
python-dateutil @ file:///home/conda/feedstock_root/build_artifacts/python-dateutil_1731919281354/work
python-nvd3==0.16.0
python-slugify==8.0.4
pytz==2024.2
pywin32==307
PyYAML==6.0.2
pyzmq @ file:///D:/bld/pyzmq_1728642354337/work
queuelib==1.7.0
referencing @ file:///home/conda/feedstock_root/build_artifacts/referencing_1733366743674/work
requests==2.32.3
requests-file==2.1.0
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rfc3339-validator==0.1.4
rich==13.9.4
rich-argparse==1.7.0
rpds-py==0.22.1
rsa==4.9
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.13.1
Scrapy==2.12.0
seaborn==0.13.2
semver==3.0.2
service-identity==24.2.0
setproctitle==1.3.4
shapely==2.0.6
six @ file:///home/conda/feedstock_root/build_artifacts/six_1620240208055/work
smmap==5.0.1
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==1.4.54
SQLAlchemy-JSONField==1.0.2
SQLAlchemy-Utils==0.41.2
sqlparse==0.5.3
stack-data @ file:///home/conda/feedstock_root/build_artifacts/stack_data_1669632077133/work
streamlit==1.19.0
tabulate==0.9.0
tenacity==9.0.0
tensorboard==2.10.1
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.10.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.5.0
text-unidecode==1.3
threadpoolctl==3.5.0
tifffile==2024.8.30
time-machine==2.16.0
tldextract==5.1.3
toml==0.10.2
tomli==2.2.1
toolz==1.0.0
tornado @ file:///D:/bld/tornado_1732615926917/work
traitlets @ file:///home/conda/feedstock_root/build_artifacts/traitlets_1713535121073/work
Twisted==24.11.0
typing_extensions @ file:///home/conda/feedstock_root/build_artifacts/typing_extensions_1717802530399/work
tzdata==2024.2
tzlocal==5.2
uc-micro-py==1.0.3
universal_pathlib==0.2.6
urllib3==2.2.3
validators==0.34.0
w3lib==2.3.1
watchdog==6.0.0
wcwidth @ file:///home/conda/feedstock_root/build_artifacts/wcwidth_1704731205417/work
Werkzeug==2.2.3
wirerope==1.0.0
wrapt==1.17.0
WTForms==3.2.1
yarl==1.18.3
zipp @ file:///home/conda/feedstock_root/build_artifacts/zipp_1732827521216/work
zope.interface==7.2
